ensure automatic selected/hovered instance settter by reacting to selected/hovered element
