+++ app/canvas/shared/use-track-hovered.ts
@@ -1,0 +1,55 @@
+import { useCallback, useEffect } from "react";
+import { type Instance, publish, useSubscribe } from "@webstudio-is/sdk";
+import {
+  useSelectedInstance,
+  useSelectedElement,
+  useHoveredElement,
+  useHoveredInstance,
+} from "./nano-states";
+import { useRootInstance } from "~/shared/nano-states";
+import { findInstanceById } from "~/shared/tree-utils";
+
+const eventOptions = {
+  passive: true,
+};
+
+export const useTrackHoveredElement = () => {
+  const [rootInstance] = useRootInstance();
+  const [, setHoveredElement] = useHoveredElement();
+
+  useEffect(() => {
+    const handleMouseOver = (event: MouseEvent) => {
+      if (rootInstance === undefined) return;
+      const element = event.target;
+      if (element instanceof HTMLElement) {
+        setHoveredElement(element);
+      }
+    };
+
+    const handleMouseOut = () => {
+      if (rootInstance === undefined) return;
+      setHoveredElement(undefined);
+    };
+
+    window.addEventListener("mouseover", handleMouseOver, eventOptions);
+    window.addEventListener("mouseout", handleMouseOut, eventOptions);
+
+    return () => {
+      window.removeEventListener("mouseover", handleMouseOver);
+      window.removeEventListener("mouseout", handleMouseOut);
+    };
+  }, [rootInstance, setHoveredElement]);
+};
+
+export const useTrackHoveredInstance = () => {
+  const [rootInstance] = useRootInstance();
+  const [hoveredElement] = useHoveredElement();
+  const [, setHoveredInstance] = useHoveredInstance();
+
+  useEffect(() => {
+    if (rootInstance === undefined || hoveredElement === undefined) return;
+    const instance = findInstanceById(rootInstance, hoveredElement.id);
+    if (instance === undefined) return;
+    setHoveredInstance(instance);
+  }, [rootInstance, hoveredElement, setHoveredInstance]);
+};
